feat(mobile): prefill continued sessions
